manageMessages bulkDelete

+++ slashCommands/moderation/bulkdelete.js
@@ -1,4 +1,4 @@
-const { EmbedBuilder, ApplicationCommandType, ActionRowBuilder, ModalBuilder, TextInputBuilder, TextInputStyle } = require('discord.js');
+const { EmbedBuilder, ApplicationCommandType, ActionRowBuilder, PermissionsBitField, ModalBuilder, TextInputBuilder, TextInputStyle } = require('discord.js');
 const Discord = require('discord.js');
 const SQLite = require("better-sqlite3");
 const bansql = new SQLite(`./databases/bans.sqlite`);
@@ -22,16 +22,22 @@ module.exports = {
     ],
     run: async (client, interaction) => {
         try {
-        const amount = interaction.options.getInteger('amount');
-        interaction.channel.messages.fetch({
-            limit: 100,
-        }).then((messages) => {
-            interaction.channel.bulkDelete(messages);
-            interaction.reply({ content: `I have deleted ${amount} messages from this channel.`, ephemeral: true })
-        });
-    } catch (error) {
-        interaction.reply({ content: `I can't deleted those messages from this channel. as they are older than 14 days`, ephemeral: true })
+            if (interaction.guild.members.me.permissions.has(PermissionsBitField.Flags.ManageMessages)) {
+                const amount = interaction.options.getInteger('amount');
+                interaction.channel.messages.fetch({
+                    limit: 100,
+                }).then((messages) => {
+                    interaction.channel.bulkDelete(messages);
+                    interaction.reply({ content: `I have deleted ${amount} messages from this channel.`, ephemeral: true })
+                });
+            } else {
+                interaction.reply({ content: `I can't deleted those messages from this channel. as I don't have permissions`, ephemeral: true })
 
-    }
-    }
+            }
+            } catch (error) {
+                interaction.reply({ content: `I can't deleted those messages from this channel. as they are older than 14 days`, ephemeral: true })
+
+            }
+
+}
 }
